pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "aiofiles>=24.1.0",
    "beautifulsoup4>=4.13.4",
    "fake-useragent>=2.2.0",
    "lxml>=5.4.0",
    "nest-asyncio>=1.6.0",
    "pillow>=11.3.0",
    "playwright>=1.54.0",
    "requests>=2.32.4",
    "trafilatura>=2.0.0",
    "urllib3>=2.5.0",
]
